# lib/auth-server/openapi.yml
openapi: 3.0.2
info:
  title: Orion authentication API
  description: Orion API for handling authentication, account creation and related activites
  contact:
    email: info@joystream.org
  license:
    name: GPL-3.0-only
    url: 'https://opensource.org/licenses/GPL-3.0'
  version: 1.0.0
servers:
  - url: '/api/v1/'
paths:
...
  /account:
    post:
      operationId: createAccount
      x-eov-operation-handler: createAccount
      description: Create a new Gateway account.
        Requires anonymousAuth to be performed first.
      security:
        - cookieAuth: []
      requestBody:
        $ref: '#/components/requestBodies/CreateAccountRequestBody'
      responses:
        '200':
          $ref: '#/components/responses/GenericOkResponse'
        '401':
          $ref: '#/components/responses/GenericUnauthorizedResponse'
        '400':
          $ref: '#/components/responses/CreateAccountBadRequestResponse'
        '404':
          $ref: '#/components/responses/CreateAccountNotFoundResponse'
        '409':
          $ref: '#/components/responses/CreateAccountConflictResponse'
        '429':
          $ref: '#/components/responses/GenericTooManyRequestsResponse'
        default:
          $ref: '#/components/responses/GenericInternalServerErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Session ID
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
  requestBodies:
...
    CreateAccountRequestBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateAccountRequestData'
...
  responses:
...
    GenericInternalServerErrorResponse:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorResponseData'
...
    CreateAccountBadRequestResponse:
      description: Cannot create user account with the provided credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorResponseData'
    GenericOkResponse:
      description: Ok
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericOkResponseData'
...
    GenericUnauthorizedResponse:
      description: Access token (session id) is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorResponseData'
...
    GenericTooManyRequestsResponse:
      description: Too many requests sent within a given timeframe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorResponseData'
...
    CreateAccountConflictResponse:
      description: Account with the provided e-mail address already exists.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorResponseData'
...
    CreateAccountNotFoundResponse:
      description: On-chain membership not found by the provided memberId.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorResponseData'
  schemas:
    ActionExecutionPayload:
      type: object
      required:
        - joystreamAccountId
        - gatewayName
        - timestamp
        - action
      properties:
        joystreamAccountId:
          type: string
          maxLength: 50
        gatewayName:
          type: string
        timestamp:
          type: number
        action:
          type: string
    ActionExecutionRequestData:
      type: object
      required:
        - payload
        - signature
      properties:
        signature:
          type: string
        payload:
          $ref: '#/components/schemas/ActionExecutionPayload'
...
    CreateAccountRequestData:
      allOf:
        - $ref: '#/components/schemas/ActionExecutionRequestData'
        - type: object
          properties:
            payload:
              allOf:
                - $ref: '#/components/schemas/ActionExecutionPayload'
                - type: object
                  required:
                    - email
                    - memberId
                  properties:
                    action:
                      type: string
                      enum:
                        - createAccount
                    memberId:
                      type: string
                    email:
                      type: string
                      pattern: ^\S+@\S+\.\S+$
                      maxLength: 255
                    encryptionArtifacts:
                      $ref: '#/components/schemas/EncryptionArtifacts'
...
    GenericErrorResponseData:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: string
    GenericOkResponseData:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
    EncryptionArtifacts:
      type: object
      required:
        - id
        - encryptedSeed
        - cipherIv
      properties:
        id:
          type: string
          maxLength: 66 # 32 bytes, hex-encoded + 2 bytes for optional 0x prefix
        encryptedSeed:
          type: string
          maxLength: 98 # 48 bytes (32-byte input + 16-byte padding), hex-encoded + 2 bytes for optional 0x prefix
        cipherIv:
          type: string
          maxLength: 34 # 16 bytes, hex-encoded + 2 bytes for optional 0x prefix
...